Removed every warning

+++ Engine/Source/inputclass.cpp
@@ -3,7 +3,6 @@
 ////////////////////////////////////////////////////////////////////////////////
 #include "../Headers/inputclass.h"
 
-
 InputClass::InputClass()
 {
 	m_directInput = 0;
@@ -11,21 +10,17 @@ InputClass::InputClass()
 	m_mouse = 0;
 }
 
-
 InputClass::InputClass(const InputClass& other)
 {
 }
-
 
 InputClass::~InputClass()
 {
 }
 
-
 bool InputClass::Initialize(HINSTANCE hinstance, HWND hwnd, int screenWidth, int screenHeight)
 {
 	HRESULT result;
-
 
 	// Store the screen size which will be used for positioning the mouse cursor.
 	m_screenWidth = screenWidth;
@@ -37,63 +32,63 @@ bool InputClass::Initialize(HINSTANCE hinstance, HWND hwnd, int screenWidth, int
 
 	// Initialize the main direct input interface.
 	result = DirectInput8Create(hinstance, DIRECTINPUT_VERSION, IID_IDirectInput8, (void**)&m_directInput, NULL);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Initialize the direct input interface for the keyboard.
 	result = m_directInput->CreateDevice(GUID_SysKeyboard, &m_keyboard, NULL);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Set the data format.  In this case since it is a keyboard we can use the predefined data format.
 	result = m_keyboard->SetDataFormat(&c_dfDIKeyboard);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Set the cooperative level of the keyboard to not share with other programs.
 	result = m_keyboard->SetCooperativeLevel(hwnd, DISCL_FOREGROUND | DISCL_EXCLUSIVE);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Now acquire the keyboard.
 	result = m_keyboard->Acquire();
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Initialize the direct input interface for the mouse.
 	result = m_directInput->CreateDevice(GUID_SysMouse, &m_mouse, NULL);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Set the data format for the mouse using the pre-defined mouse data format.
 	result = m_mouse->SetDataFormat(&c_dfDIMouse);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Set the cooperative level of the mouse to share with other programs.
 	result = m_mouse->SetCooperativeLevel(hwnd, DISCL_FOREGROUND | DISCL_NONEXCLUSIVE);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
 
 	// Acquire the mouse.
 	result = m_mouse->Acquire();
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		return false;
 	}
@@ -101,11 +96,10 @@ bool InputClass::Initialize(HINSTANCE hinstance, HWND hwnd, int screenWidth, int
 	return true;
 }
 
-
 void InputClass::Shutdown()
 {
 	// Release the mouse.
-	if(m_mouse)
+	if (m_mouse)
 	{
 		m_mouse->Unacquire();
 		m_mouse->Release();
@@ -113,7 +107,7 @@ void InputClass::Shutdown()
 	}
 
 	// Release the keyboard.
-	if(m_keyboard)
+	if (m_keyboard)
 	{
 		m_keyboard->Unacquire();
 		m_keyboard->Release();
@@ -121,7 +115,7 @@ void InputClass::Shutdown()
 	}
 
 	// Release the main interface to direct input.
-	if(m_directInput)
+	if (m_directInput)
 	{
 		m_directInput->Release();
 		m_directInput = 0;
@@ -130,22 +124,20 @@ void InputClass::Shutdown()
 	return;
 }
 
-
 bool InputClass::Frame()
 {
 	bool result;
 
-
 	// Read the current state of the keyboard.
 	result = ReadKeyboard();
-	if(!result)
+	if (!result)
 	{
 		return false;
 	}
 
 	// Read the current state of the mouse.
 	result = ReadMouse();
-	if(!result)
+	if (!result)
 	{
 		return false;
 	}
@@ -156,18 +148,16 @@ bool InputClass::Frame()
 	return true;
 }
 
-
 bool InputClass::ReadKeyboard()
 {
 	HRESULT result;
 
-
 	// Read the keyboard device.
 	result = m_keyboard->GetDeviceState(sizeof(m_keyboardState), (LPVOID)&m_keyboardState);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		// If the keyboard lost focus or was not acquired then try to get control back.
-		if((result == DIERR_INPUTLOST) || (result == DIERR_NOTACQUIRED))
+		if ((result == DIERR_INPUTLOST) || (result == DIERR_NOTACQUIRED))
 		{
 			m_keyboard->Acquire();
 		}
@@ -176,22 +166,20 @@ bool InputClass::ReadKeyboard()
 			return false;
 		}
 	}
-		
+
 	return true;
 }
-
 
 bool InputClass::ReadMouse()
 {
 	HRESULT result;
 
-
 	// Read the mouse device.
 	result = m_mouse->GetDeviceState(sizeof(DIMOUSESTATE), (LPVOID)&m_mouseState);
-	if(FAILED(result))
+	if (FAILED(result))
 	{
 		// If the mouse lost focus or was not acquired then try to get control back.
-		if((result == DIERR_INPUTLOST) || (result == DIERR_NOTACQUIRED))
+		if ((result == DIERR_INPUTLOST) || (result == DIERR_NOTACQUIRED))
 		{
 			m_mouse->Acquire();
 		}
@@ -204,7 +192,6 @@ bool InputClass::ReadMouse()
 	return true;
 }
 
-
 void InputClass::ProcessInput()
 {
 	// Update the location of the mouse cursor based on the change of the mouse location during the frame.
@@ -212,20 +199,19 @@ void InputClass::ProcessInput()
 	m_mouseY += m_mouseState.lY;
 
 	// Ensure the mouse location doesn't exceed the screen width or height.
-	if(m_mouseX < 0)  { m_mouseX = 0; }
-	if(m_mouseY < 0)  { m_mouseY = 0; }
-	
-	if(m_mouseX > m_screenWidth)  { m_mouseX = m_screenWidth; }
-	if(m_mouseY > m_screenHeight) { m_mouseY = m_screenHeight; }
-	
+	if (m_mouseX < 0) { m_mouseX = 0; }
+	if (m_mouseY < 0) { m_mouseY = 0; }
+
+	if (m_mouseX > static_cast<float>(m_screenWidth)) { m_mouseX = static_cast<float>(m_screenWidth); }
+	if (m_mouseY > static_cast<float>(m_screenHeight)) { m_mouseY = static_cast<float>(m_screenHeight); }
+
 	return;
 }
-
 
 bool InputClass::IsEscapePressed()
 {
 	// Do a bitwise and on the keyboard state to check if the escape key is currently being pressed.
-	if(m_keyboardState[DIK_ESCAPE] & 0x80)
+	if (m_keyboardState[DIK_ESCAPE] & 0x80)
 	{
 		return true;
 	}
@@ -244,10 +230,9 @@ bool InputClass::KeyIsPressed(unsigned char DI_keycode)
 	return false;
 }
 
-
 void InputClass::GetMouseLocation(int& mouseX, int& mouseY)
 {
-	mouseX = m_mouseX;
-	mouseY = m_mouseY;
+	mouseX = static_cast<int>(m_mouseX);
+	mouseY = static_cast<int>(m_mouseY);
 	return;
 }
